backend/pyproject.toml
[tool.poetry]
name = "workout-log"
version = "0.1.0"
description = "An application to track your workouts"
authors = ["Guillaume <7831777+GuillaumeOj@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
fastapi = {extras = ["all"], version = "*"}
sqlmodel = "^0.0.14"
email-validator = "^2.1.0.post1"
greenlet = "^3.0.3"
alembic = "^1.13.1"
sqlalchemy = {version = "^2.0.25"}
psycopg = {extras = ["binary"], version = "^3.1.17"}
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
pydantic-settings = "^2.2.1"
daiquiri = "^3.2.5.1"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.12.0"
flake8 = "^7.0.0"
mypy = "<1.9.0"
pdbpp = "^0.10.3"
ipython = "^8.20.0"
httpx = "^0.26.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.coverage.run]
omit = ["app/tests/*"]
